requirements.txt
discord.py==1.7.3
python-dotenv==0.19.0
requests==2.25.1